moved ClientCapabilityProviderTests to it's own file
